Comentarios y cambio de logica
